alugueres visiveis

+++ queries/cars/delete_car.php
@@ -1,6 +1,10 @@
 <?php 
     include '../../connection/connection.php';
 
+    if (!isset($_SESSION['logged_in']) || $_SESSION['logged_in'] !== true) {
+        header("Location: login.php");
+        exit; 
+    }
     
     $sql = "DELETE FROM cars WHERE id='$id'";
